% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerPyramid.R
\name{acgtPyramid}
\alias{acgtPyramid}
\title{This is a test function}
\usage{
acgtPyramid(acgtDistributionList, colorList, ids = NULL, show.edges = FALSE,
  text = NULL, text.cex = 1, show.identify = FALSE, method = "scatter")
}
\arguments{
\item{acgtDistributionList}{distribution of the acgt}

\item{colorList}{list of color values}

\item{ids}{names of samples or species}

\item{show.edges}{should the deges in the scatter plot shown? (default = FALSE)}

\item{method}{should a scatterplot (scatter) or 3d plot (3d) be plotted (default = scatter)}
}
\description{
This is a test function
}
\details{
This is a test function
}
\author{
Jochen Kruppa
}

