% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacteriaKmerDistribution}
\alias{bacteriaKmerDistribution}
\title{k-mer distribution of bacteria DNA sequences}
\format{A data frame with 5785 rows and 1365 variables:
\describe{
  \item{id}{Name of the virus}
  \item{A}{Percentage of nucleic acid base A in the 1-mers}
  \item{...}{...}
  \item{TTTTT}{Percentage of nucleic acid mer TTTTT in the 5-mers}
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC99098/}
}
\usage{
bacteriaKmerDistribution
}
\description{
A dataset containing containing the 1, 2, 3, 4, and 5-mer
distribution of the nucleic acid bases of 5785 viruses available as
embl files from NCBI.
}
\keyword{datasets}

