% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerPyramid.R
\name{baseFunction}
\alias{baseFunction}
\title{This is a test function}
\usage{
baseFunction(foo, bar)
}
\arguments{
\item{foo}{}

\item{bar}{}
}
\description{
This is a test function
}
\details{
This is a test function
}
\author{
Jochen Kruppa
}

