% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerDistribution.R
\name{generatePairedReads}
\alias{generatePairedReads}
\title{Generate paired reads from a given genome}
\usage{
generatePairedReads(numReads = 10, initSequence, readLength = 150,
  readMeanDist = 50, files, parallel_flag = FALSE)
}
\arguments{
\item{numReads}{Number of wanted reads}

\item{initSequence}{Original genome sequence to draw reads from}

\item{readLength}{Wanted read length in bases}

\item{readMeanDist}{Mean distance between the paired reads
(default = 50)}

\item{files}{Vector of two named files ("R1", "R2") to save the
generated reads}

\item{parallel_flag}{Should everything done in parallel (not
recommended).}
}
\description{
The function generates two files with paired reads
}
\details{
The generated files (R1 and R2) are generated by drawing the
wanted reads from a given initial genome sequence
}
\author{
Jochen Kruppa
}

