% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_pyramid_3d.R
\name{get_diff_values}
\alias{get_diff_values}
\title{Get the difference in counts between to sequences}
\usage{
get_diff_values(list)
}
\arguments{
\item{list}{}
}
\description{
Get the difference in counts between to sequences
}
\details{
Get the difference in counts between to sequences
}
\author{
Jochen Kruppa
}

