% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_pyramid_3d.R
\name{get_pca_plot_list}
\alias{get_pca_plot_list}
\title{Calculate the PCA with pyramid edges}
\usage{
get_pca_plot_list(df)
}
\arguments{
\item{df}{}
}
\description{
Calculate the PCA with pyramid edges
}
\details{
Calculate the PCA with pyramid edges
}
\author{
Jochen Kruppa
}

