% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_pyramid_3d.R
\name{get_pca_window_list}
\alias{get_pca_window_list}
\title{Get the PCA values for sliding window}
\usage{
get_pca_window_list(seqs, window)
}
\arguments{
\item{seqs}{}

\item{window}{}
}
\description{
Get the PCA values for sliding window
}
\details{
Get the PCA values for sliding window
}
\examples{
data(viralExampleSeqs)

viral_window_list <- get_pca_window_list(viralExampleSeqs, window = 2)
}
\author{
Jochen Kruppa
}

