% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_distribution.R
\name{get_reads_num}
\alias{get_reads_num}
\title{Calculate number of reads given the coverage and genome
    size}
\usage{
get_reads_num(coverage, readLength, genomeSize)
}
\arguments{
\item{coverage}{Wanted coverage}

\item{readLength}{Given read length}

\item{genomeSize}{Size of the genome in bases}
}
\value{
numeric, number of reads
}
\description{
The function calculates the needed reads given a coverage and the
genome size
}
\details{
The calulation is done by the formula of Lander and Waterman (1988).
}
\references{
Lander, E. S. and Waterman, M. S. (1988). Genomic
    mapping by fingerprinting random clones: A mathematical
    analysis. Genomics, 2(3), 231-239.
}
\author{
Jochen Kruppa
}
