% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_pyramid_3d.R
\name{get_window_distr}
\alias{get_window_distr}
\title{Wrapper for the sliding window k-mer distribution}
\usage{
get_window_distr(seq, window, prob_flag = TRUE)
}
\arguments{
\item{seq}{}

\item{window}{}

\item{prob_flag}{}
}
\description{
Wrapper for the sliding window k-mer distribution
}
\details{
Wrapper for the sliding window k-mer distribution
}
\author{
Jochen Kruppa
}

