% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerDistribution.R
\name{getReadsNum}
\alias{getReadsNum}
\title{Calculate number of reads given the coverage and genome size}
\usage{
getReadsNum(coverage, readLength, genomeSize)
}
\arguments{
\item{coverage}{Wanted coverage}

\item{readLength}{Given read length}

\item{genomeSize}{Size of the genome in bases}
}
\value{
numeric, number of reads
}
\description{
Text description
}
\details{
Details
}
\author{
Jochen Kruppa
}

