% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerPyramid.R
\name{kmerPyramid}
\alias{kmerPyramid}
\title{This is a test function}
\usage{
kmerPyramid(kmerDistributionList, colorList, ids = NULL, show.edges = FALSE,
  method = "scatter")
}
\arguments{
\item{kmerDistributionList}{}

\item{colorList}{}

\item{ids}{}

\item{show.edges}{}

\item{method}{}
}
\description{
This is a test function
}
\details{
This is a test function
}
\author{
Jochen Kruppa
}

