% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_pyramid_3d.R
\name{pyramid_3d}
\alias{pyramid_3d}
\title{ACGT pyramid 3D plotting function}
\usage{
pyramid_3d(df, color = "black", ids = NULL, text = NULL, cex = 1,
  identify = FALSE)
}
\arguments{
\item{df}{Data frame of k-mer freuencies. The single k-mers are
the columns and the rows indicating different samples, sequence
reads, or contigs.}

\item{color}{Single value 'black', if all points should be black
or vector of length \emph{nrow(df)} if all points should be
colored differently.}

\item{ids}{If \emph{identify = TRUE} the single points can be
selected and the given id is shown. Must have the length
\emph{nrow(df)}.}

\item{text}{Single value 'x', if all points should be printed as
'x' or vector of length \emph{nrow(df)} if all points should be
printed by a different letter differently.}

\item{cex}{Size of the text.}

\item{identify}{Set to TRUE, if points should be identified by
their \empgh{ids}.}
}
\description{
The acgt pyramid 3D plotting function
}
\details{
The acgt pyramid 3D plotting function
}
\examples{
data(viralExampleSeqs)

kmer_distr <- getKmerDistribution(viralExampleSeqs, k = 1)

pyramid_3d(kmer_distr,
           cex = 2,
           color = "blue")

ids <- names(viralExampleSeqs)

pyramid_3d(kmer_distr,
           ids = ids,
           cex = 2,
           color = "blue",
           identify = TRUE)

data(viralExampleCodingSeq)

kmer_distr <- getKmerDistribution(viralExampleCodingSeq, k = 1)
text_ids <- ifelse(names(viralExampleCodingSeq) == "non_coding", "x", "o")
color_ids <- ifelse(names(viralExampleCodingSeq) == "non_coding", "black", "red")

pyramid_3d(kmer_distr,
           cex = 1,
           text = text_ids,
           color = color_ids)

ids <- names(viralExampleCodingSeq)

pyramid_3d(kmer_distr,
           ids = ids,
           cex = 1,
           text = text_ids,
           color = color_ids,
           identify = TRUE)
}
\author{
Jochen Kruppa
}

