% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_pyramid_3d.R
\name{pyramid_3d_draw_edges}
\alias{pyramid_3d_draw_edges}
\title{Wrapper for getting the positions of the edges}
\usage{
pyramid_3d_draw_edges(pca_edge_df, alpha = 1)
}
\arguments{
\item{pca_edge_df}{}
}
\description{
Wrapper for getting the positions of the edges
}
\details{
Wrapper for getting the positions of the edges
}
\author{
Jochen Kruppa
}

