% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_pyramid_3d.R
\name{pyramid_3d_window}
\alias{pyramid_3d_window}
\title{The k-mer window pyramid 3D plotting function}
\usage{
pyramid_3d_window(list, color = "black", difference = FALSE,
  identify = FALSE, bw = FALSE, bw.cex = 10, edges = NULL)
}
\arguments{
\item{list}{List generated of sequences by the function
\code{link{get_pca_window_list}}. Only one list entry can be
plotted at the same time. If two list entries are provided, set
difference to TRUE to get the difference pyramid.}

\item{color}{Single value 'black', if all points should be black
or vector of length \emph{nrow(df)} if all points should be
colored differently.}

\item{difference}{Should the difference between two list entries
be calcualted and the difference pyramid be plotted? [default =
FALSE]}

\item{identify}{Set to TRUE, if points should be identified, than
the k-mer will be shown.}

\item{bw}{Black and white plot for the differences [default =
FALSE]}

\item{bw.cex}{Size of the text symbols in the black and white
difference plot.}

\item{edges}{A vector of edges which should be printed}
}
\description{
The k-mer window pyramid 3D plotting function
}
\details{
The k-mer window pyramid 3D plotting function
}
\examples{
data(viralExampleSeqs)

viral_window_list <- get_pca_window_list(viralExampleSeqs, window = 2)

pyramid_3d_window(viral_window_list[1],
                  color = "red")

pyramid_3d_window(viral_window_list[1],
                  color = "red",
                  identify = TRUE)

pyramid_3d_window(viral_window_list[c(3,5)],
                  difference = TRUE,
                  identify = TRUE)

pyramid_3d_window(viral_window_list[c(3,5)],
                  difference = TRUE,
                  bw = TRUE,
                  bw.cex = 75,
                  identify = TRUE)

pyramid_3d_window(viral_window_list[1],
                  color = "red",
                  edges = c("AT", "CG"))
}
\author{
Jochen Kruppa
}

