% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_to_null.R
\name{rotate_to_null}
\alias{rotate_to_null}
\title{Experimental function for rotating the PCA coordinates}
\usage{
rotate_to_null(list)
}
\arguments{
\item{list}{}
}
\value{
List of rotated PCA coordinates
}
\description{
Experimental function for rotating the PCA coordinates
}
\details{
Experimental function for rotating the PCA coordinates
}
\examples{

### edges
A <- c(1, 1, 1)
C <- c(1, 1, 2)
G <- c(1, 2, 1)
T <- c(2, 1, 1)
### Further points
X <- matrix(rnorm(20*3), 20, 3)
rotate_to_null(A, C, G, T, X)
}
\author{
Jochen Kruppa
}

