% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerPyramid.R
\name{scatterplot3d.invisible}
\alias{scatterplot3d.invisible}
\title{Small helper function to make a invisible scatterplot3d}
\usage{
scatterplot3d.invisible(...)
}
\arguments{
\item{...}{}
}
\value{
scatterplot3d object
}
\description{
Small helper function to make a invisible scatterplot3d
}
\details{
Small helper function to make a invisible scatterplot3d
}
\author{
Jochen Kruppa
}

