% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{translate.phred}
\alias{translate.phred}
\title{Test title for later}
\usage{
translate.phred(QS, plot = TRUE, title = "")
}
\arguments{
\item{QS}{}

\item{plot}{}

\item{title}{}
}
\description{
Test title for later
}
\details{
Test title for later
}
\author{
Klaus Jung
}
