% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viralExampleCodingSeq}
\alias{viralExampleCodingSeq}
\title{DNAStringSet of coding and non coding of NC_015230.1 viral strain}
\format{A DNAStringSet with 41 entries of different width
\describe{
  \item{width}{Length of the viral sequence}
  \item{seq}{DNA sequence of the viral sequence}
  \item{names}{NCBI genbank ID and dsecription if CDS otherwise names non_coding}
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/viral/}
}
\usage{
viralExampleCodingSeq
}
\description{
A DNAStringSet containing coding and non coding sequences of the
NC_015230.1 viral strain
}
\keyword{datasets}

