% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viralExampleSeqs}
\alias{viralExampleSeqs}
\title{DNAStringSet of ten viral sequences}
\format{A DNAStringSet with 10 entries of different width
\describe{
  \item{width}{Length of the viral sequence}
  \item{seq}{DNA sequence of the viral sequence}
  \item{names}{NCBI genbank ID and description}
}}
\source{
\url{ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/viral/}
}
\usage{
viralExampleSeqs
}
\description{
A DNAStringSet containing ten randomly choosen viral sequences for
example purpose.
}
\keyword{datasets}

