% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diff_values.R
\name{get_diff_values}
\alias{get_diff_values}
\title{Get the difference in counts between to sequences}
\usage{
get_diff_values(list)
}
\arguments{
\item{list}{generated by \code{\link{get_pca_window_list}}}
}
\value{
data.frame including the PCA coordinates and the radius
  differences
}
\description{
This is an internal function, which uses the results from the
function \code{\link{get_pca_window_list}} to generate the
differences in the k-mer distribution between two samples.
}
\details{
This is an internal function.
}
\author{
Jochen Kruppa
}

