% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_window_distr.R
\name{get_window_distr}
\alias{get_window_distr}
\title{Wrapper for the sliding window k-mer distribution}
\usage{
get_window_distr(seq, window, prob_flag = TRUE)
}
\arguments{
\item{seq}{DNAStringSet object}

\item{window}{Size of the window}

\item{prob_flag}{Should the frequencies of the k-mers returned or the counts}
}
\value{
Data frame of the k-mer distribution.
}
\description{
Wrapper for the sliding window k-mer distribution
}
\details{
Wrapper for the sliding window k-mer distribution
}
\author{
Jochen Kruppa
}

