% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_distribution.R
\name{getKmerDistribution}
\alias{getKmerDistribution}
\title{Get the k-mer distribution from a sequence}
\usage{
getKmerDistribution(seq, kmers = c(1, 2, 3, 4, 5))
}
\arguments{
\item{seq}{Sequence of characters}

\item{kmers}{vector of wanted k-mers}
}
\value{
data.frame with probability of the k-mers
}
\description{
Calculates the k-mer distribution of a given sequence
}
\details{
Calculates the k-mer distribution of a given sequence
}
\author{
Jochen Kruppa
}

