% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerPyramid.R
\name{kmerPyramid}
\alias{kmerPyramid}
\title{This is a test function}
\usage{
kmerPyramid(sampleKmerDistribution, virusKmerDistribution = NULL,
  edges = FALSE, colorVirus = "yellow", colorSample = NULL)
}
\arguments{
\item{sampleKmerDistribution}{}

\item{virusKmerDistribution}{}

\item{edges}{}

\item{colorVirus}{}

\item{colorSample}{}
}
\description{
This is a test function
}
\details{
This is a test function
}
\author{
Jochen Kruppa
}

