% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid_3d_draw_edges.R
\name{pyramid_3d_draw_edges}
\alias{pyramid_3d_draw_edges}
\title{Function to draw the edges of the 3D pyramid}
\usage{
pyramid_3d_draw_edges(pca_edge_df, alpha = 1)
}
\arguments{
\item{pca_edge_df}{Data frame created by the
\code{\link{get_pca_plot_list}} function for drawing the edges
of the 3D pyramid.}
}
\description{
Function to draw the edges of the 3D pyramid
}
\details{
Function to draw the edges of the 3D pyramid
}
\author{
Jochen Kruppa
}
