% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid_3d_window.R
\name{pyramid_3d_window}
\alias{pyramid_3d_window}
\title{The k-mer window pyramid 3D plotting function}
\usage{
pyramid_3d_window(list, color = "black", difference = FALSE,
  identify = FALSE, bw = FALSE, bw.cex = 10, edges = NULL)
}
\arguments{
\item{list}{List of sequences generated by the function
\code{link{get_pca_window_list}}. Only one list entry can be
plotted at the same time. If two list entries are provided, set
difference to TRUE to get the difference pyramid.}

\item{color}{Single value 'black', if all points should be black
or vector of length \emph{nrow(df)} if all points should be
colored differently.}

\item{difference}{Should the difference between two list entries
be calcualted and the difference pyramid be plotted? [default =
FALSE]}

\item{identify}{Set to TRUE, if points should be identified, than
the k-mer will be shown.}

\item{bw}{Black and white plot for the differences [default =
FALSE]}

\item{bw.cex}{Size of the text symbols in the black and white
difference plot.}

\item{edges}{A vector of edges which should be printed. Because
there is no difference between AG and GA only the alpha numeric
one will work: AG.}
}
\description{
The k-mer window pyramid 3D plotting function allows to plot the
k-mer distribution given by a window of a single sequences. Further,
it is possible to compare two sequences to each other.
}
\details{
The function plots the k-mer distribution of single sequences
generated by the function \code{link{get_pca_window_list}} into
the three dimensional space of a principal component analysis.

The user can provide single list entries or two list entries. In
the first case the k-mer distribution given by the window size
will be plotted. The color and can be changed by \emph{color} or
different edges can be shown by \emph{edges}. If two list entries
are supplied and \emph{difference} is set to TRUE the difference
in both sequences given the window k-mer count is calculated and
shown by different radius sizes. An increase from sequence one to
sequence two is colored blue and a decrease is colored red. The
identify option allows to identify single points in the 3D plot.
}
\examples{
## Read in own DNA sequences by the package Biostrings (see Details for more information)

data(viralExampleSeqs)

viral_window_list <- get_pca_window_list(viralExampleSeqs, window = 2)

pyramid_3d_window(viral_window_list[1],
                  color = "red")

pyramid_3d_window(viral_window_list[1],
                  color = "red",
                  identify = TRUE)

pyramid_3d_window(viral_window_list[c(3,5)],
                  difference = TRUE,
                  identify = TRUE)

pyramid_3d_window(viral_window_list[c(3,5)],
                  difference = TRUE,
                  bw = TRUE,
                  bw.cex = 75,
                  identify = TRUE)

pyramid_3d_window(viral_window_list[1],
                  color = "red",
                  edges = c("AT", "CG"))
}
\author{
Jochen Kruppa
}
