% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_pyramid_3d.R
\name{pyramid_3d_window}
\alias{pyramid_3d_window}
\title{The k-mer window pyramid 3D plotting function}
\usage{
pyramid_3d_window(list, color = "black", difference = FALSE,
  identify = FALSE)
}
\arguments{
\item{list}{}

\item{color}{}

\item{difference}{}

\item{identify}{}
}
\description{
The k-mer window pyramid 3D plotting function
}
\details{
The k-mer window pyramid 3D plotting function
}
\examples{
load("../data/viralExampleSeqs.rda")

viral_window_list <- get_pca_window_list(viralExampleSeqs, window = 2)

pyramid_3d_window(viral_window_list[1],
                  color = "red")

pyramid_3d_window(viral_window_list[1],
                  color = "red",
                  identify = TRUE)

pyramid_3d_window(viral_window_list[c(3,5)],
                  difference = TRUE)

pyramid_3d_window(viral_window_list[c(3,5)],
                  difference = TRUE,
                  identify = TRUE)
}
\author{
Jochen Kruppa
}

