% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{read.fastq}
\alias{read.fastq}
\title{Test title for later}
\usage{
read.fastq(path, dat, nlines = 0)
}
\arguments{
\item{path}{}

\item{dat}{}

\item{nlines}{}
}
\description{
Test title for later
}
\details{
Test title for later
}
\author{
Klaus Jung
}

